refactor(runtime_ipc): split tests to separate file (#1266)
Splits src/runtime_ipc.rs (406 LOC) into:
- src/runtime_ipc/mod.rs   (~248 LOC) — production code
- src/runtime_ipc/tests.rs (~158 LOC) — test module body

Pure refactor: zero logic changes, all 12 runtime_ipc tests pass unmodified.
Brings module under 400-line rule.

Closes 1/70 of #1266 (after #1285, #1286).
